Fix bug that rendered incorrect tokens
Instead of rendering the token at the index on which move was played,
the last token in gameboard was being printed, but it stores the tokens
at indices they were played on. So if the gameboard contained "X"
as last element at Index 4, and then O is played at index 2,
"O" will get stored at gameboard[2] but "X" will be rendered out because
the last element is still X.
Now the element is rendered based on the index that was played.